[backstop] refact: rename entry file for package
