--- packages/visual-regression-testing/lib/visual-regression-testing.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = visualRegressionTesting;
-
-function visualRegressionTesting() {
-    // TODO
-}
