+++ packages/visual-regression-testing/package.json
@@ -5,7 +5,7 @@
   "author": "laras126 <laras126@gmail.com>",
   "homepage": "https://github.com/penske-media-corp/pmc-larva#readme",
   "license": "ISC",
-  "main": "lib/visual-regression-testing.js",
+  "main": "index.js",
   "directories": {
     "lib": "lib",
     "test": "__tests__"
